netshop/requirements.txt
Django==2.2.5
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.8.0
itsdangerous==1.1.0
jieba==0.39
jsonpickle==1.2
mysqlclient==1.4.4
Pillow==6.1.0
pycryptodomex==3.7.2
python-alipay-sdk==1.10.1
pytz==2019.2
redis==3.3.8
sqlparse==0.3.0
Whoosh==2.7.4
